modification in C02 ex06 to ex10

+++ pool/C01/ex06/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: tbourill <terry.bourillon@outlook.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/19 05:49:52 by tbourill          #+#    #+#             */
-/*   Updated: 2024/07/19 08:24:23 by tbourill         ###   ########.fr       */
+/*   Updated: 2024/07/23 03:46:12 by tbourill         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-size_t	ft_strlen(char*txt)
+unsigned int	ft_strlen(char*txt)
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (txt[i] != '\0')
